requirements.txt
pymongo~=3.12.0
python-dotenv~=1.1.1
pytest

mongomock~=4.3.0